requirements.txt
asgiref==3.9.2
attrs==25.3.0
black==25.9.0
cfgv==3.4.0
click==8.3.0
distlib==0.4.0
Django==5.2.7
django-environ==0.12.0
django-filter==25.1
djangorestframework==3.16.1
drf-spectacular==0.28.0
factory_boy==3.3.3
Faker==37.8.0
filelock==3.19.1
identify==2.6.14
inflection==0.5.1
iniconfig==2.1.0
isort==6.0.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
mypy_extensions==1.1.0
nodeenv==1.9.1
packaging==25.0
pathspec==0.12.1
platformdirs==4.4.0
pluggy==1.6.0
pre_commit==4.3.0
psycopg==3.2.10
psycopg-binary==3.2.10
psycopg2==2.9.10
Pygments==2.19.2
pytest==8.4.2
pytest-django==4.11.1
pytokens==0.1.10
PyYAML==6.0.3
referencing==0.36.2
rpds-py==0.27.1
ruff==0.13.2
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
virtualenv==20.34.0
